refactor: move classes related to common
