+++ backend/core/src/main/java/com/scoinone/core/common/annotation/LoginUser.java
@@ -1,4 +1,4 @@
-package com.scoinone.core.auth;
+package com.scoinone.core.common.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
