+++ backend/core/src/main/java/com/scoinone/core/controller/BuyOrderController.java
@@ -1,6 +1,6 @@
 package com.scoinone.core.controller;
 
-import com.scoinone.core.auth.LoginUser;
+import com.scoinone.core.common.annotation.LoginUser;
 import com.scoinone.core.dto.common.DeleteResponseDto;
 import com.scoinone.core.dto.request.order.CreateBuyOrderRequestDto;
 import com.scoinone.core.dto.response.order.CreateBuyOrderResponseDto;
