+++ backend/core/src/main/java/com/scoinone/core/service/impl/TradeServiceImpl.java
@@ -1,6 +1,6 @@
 package com.scoinone.core.service.impl;
 
-import com.scoinone.core.common.OrderStatus;
+import com.scoinone.core.common.status.OrderStatus;
 import com.scoinone.core.entity.BuyOrder;
 import com.scoinone.core.entity.OwnedVirtualAsset;
 import com.scoinone.core.entity.SellOrder;
