+++ backend/core/src/test/java/com/scoinone/core/integration/repository/BuyOrderRepositoryTest.java
@@ -2,7 +2,7 @@ package com.scoinone.core.integration.repository;
 
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import com.scoinone.core.common.OrderStatus;
+import com.scoinone.core.common.status.OrderStatus;
 import com.scoinone.core.config.TestContainerConfig;
 import com.scoinone.core.entity.BuyOrder;
 import com.scoinone.core.entity.User;
